Add files via upload
Random number generator that's modified to run until it finds a number that is on a list. The list is reference numbers for different dinosaur genera, from several different taxa I'm interested in. (Yes, I know this is only minimally useful, and I could have saved a lot of time by just picking one species instead of writing this)